bugfix(#170):localstorage 오류 해결
